Why multiple examples are VITAL to help learning via patterns
A growing documentation library, with MORE that one example to explain
each concept or language feature.  If someone gives you only on
example, you have to rely on their written explanation.  If someone
gives you two examples, you can start to look for things that are part
of the language syntax & things that have named by the user.  Without
at least two examples, it is impossible to look for patterns in the
language.  I really think this is something which is NEEDED.